Added tile editor
(heavy programmer art, so it'll be updated again in the future, but the basic idea is there. top left is for palettes, bottom left for collisions (colours' meanings are in the useful_things spreadsheet), top right for selecting a subtile, and the changing menu in the middle left is for editing subtiles and constructing bigtile graphics/collisions. the four buttons from left to right are for priority (should the last 3 colours in the palette be drawn over the bears), vertical flip, horizontal flip, and bank (change from the first 255 subtiles to the next 255).)